+++ core/src/main/java/com/mawkun/core/service/InvestLogServiceExt.java
@@ -31,7 +31,7 @@ public class InvestLogServiceExt extends InvestLogService {
         return new PageInfo<>(list);
     }
 
-    public int save(User user, MemberCard cart, Integer cartNum, Double investMoney, Double giftMoney, Double amoutMoney, Double residueMoney) {
+    public int save(User user, MemberCard cart, Long cartNum, Long investMoney, Long giftMoney, Long amoutMoney, Long residueMoney) {
         InvestLog investLog = new InvestLog();
         investLog.setUserId(user.getId());
         investLog.setUserName(user.getUserName());
